Moves TestTemplateStrings to correct location of template_test.go
